LU-17084 lod: fix comparision in lod_striping_load()
in if (rc > sizeof(struct lmv_foreign_md)) the latter
is unsigned and gcc treats rc (which is defined as int
and can be negative to encode an error) as unsigned.
this way -EIO becomes greater than the size of the
structure.  make sizeof() signed to avoid confusion.

Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/52265
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/lod/lod_lov.c
@@ -1629,7 +1629,8 @@ int lod_striping_load(const struct lu_env *env, struct lod_object *lo)
 			lo->ldo_comp_cached = 1;
 	} else if (S_ISDIR(lod2lu_obj(lo)->lo_header->loh_attr)) {
 		rc = lod_get_lmv_ea(env, lo);
-		if (rc > sizeof(struct lmv_foreign_md)) {
+
+		if (rc > (int)sizeof(struct lmv_foreign_md)) {
 			struct lmv_foreign_md *lfm = info->lti_ea_store;
 
 			if (le32_to_cpu(lfm->lfm_magic) == LMV_MAGIC_FOREIGN) {
